Split dtls test.
Move correlation tests into separate test class.
Use ConnectorHelper for the correlation tests.
